% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histo.R
\name{histo}
\alias{histo}
\title{Histogram of sample means}
\usage{
histo(n, iter)
}
\arguments{
\item{n}{- number}

\item{iter}{- iterations}
}
\description{
Histogram of sample means
}
\examples{
histo(n=10,iter=10000)
}
