% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearscatter.R
\name{linearscatter}
\alias{linearscatter}
\title{Create a simple plot with a linear regression line}
\usage{
linearscatter(df, y, x, xlb, ylb)
}
\arguments{
\item{df}{Data table}

\item{y}{Dependent variable name}

\item{x}{Independent variable name}

\item{xlb}{x-axis label}

\item{ylb}{y-axis label}
}
\value{
A plot with independent and dependent variables labelled with a linear regression line.
}
\description{
Create a simple plot with a linear regression line
}
\examples{
spruce.df <- read.csv("SPRUCE.csv")
linearscatter(spruce.df, "Height", "BHDiameter", "Breast Height Diameter (cm)", "Height (m)")
}
