% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybins.R
\name{mybin}
\alias{mybin}
\title{Makes a Normal Through Bins}
\usage{
mybin(iter = 100, n = 10, p = 0.7)
}
\arguments{
\item{iter}{- iterations}

\item{n}{Sample size}

\item{p}{- probability}
}
\value{
A histogram
}
\description{
Creates a distribution graph of a random sample
}
\examples{
mybin(n=1000)
}
