% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot2.R
\name{myboot2}
\alias{myboot2}
\title{myBoot function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{- iterations}

\item{x}{- data}

\item{fun}{- mean}

\item{alpha}{- 0.05}

\item{...}{- ...}
}
\description{
myBoot function
}
\examples{
myboot2(x=(MATH4753PRESTONROESSLETFALL23::fire)$DAMAGE)
}
