% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My N Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{- mu}

\item{sigma}{- sigma}

\item{a}{- ad}
}
\value{
The area (probability, P(X<=a))
}
\description{
My N Curve
}
