% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCSV.R
\name{myRead}
\alias{myRead}
\title{A function to read in a csv file}
\usage{
myRead(csv)
}
\arguments{
\item{csv}{A csv file}
}
\value{
A dataframe
}
\description{
A function to read in a csv file
}
\examples{
myRead <- ("SPRUCE.csv")
}
