% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{NTickets Function
This method calculates the number of tickets to be sold
using discrete distribution and normal approximation.}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{- number of seats available on the plane}

\item{gamma}{- probability the plane will be overbooked}

\item{p}{- probability that a given person will show up}
}
\description{
NTickets Function
This method calculates the number of tickets to be sold
using discrete distribution and normal approximation.
}
\examples{
ntickets(400, 0.02, 0.95)
}
